LayoutTests/imported/w3c/web-platform-tests/content-security-policy/securitypolicyviolation/source-file-blob-scheme-expected.txt


FAIL Violations from data:-URL scripts have a sourceFile of 'blob' assert_equals: expected 17 but got 21
